setup preliminary graphql schema

+++ kotlin-core/src/main/resources/db/schema.sql
@@ -6,18 +6,18 @@ DROP TABLE IF EXISTS users;
 DROP TABLE IF EXISTS groups;
 
 CREATE TABLE groups (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     group_name text NOT NULL
 );
 
 CREATE TABLE users (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     username TEXT NOT NULL
 );
 
 CREATE TABLE users_groups (
-    group_id INTEGER,
-    user_id INTEGER,
+    group_id BIGINT,
+    user_id BIGINT,
     CONSTRAINT pk_user_groups
         PRIMARY KEY(group_id, user_id),
     CONSTRAINT fk_user_groups_group_id
@@ -29,20 +29,20 @@ CREATE TABLE users_groups (
 );
 
 CREATE TABLE categories (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     title TEXT NOT NULL,
-    group_id INTEGER,
+    group_id BIGINT,
     CONSTRAINT fk_categories_group_id
         FOREIGN KEY(group_id)
             REFERENCES groups(id) ON DELETE CASCADE
 );
 
 CREATE TABLE discussions (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     heading TEXT NOT NULL,
     sub TEXT,
-    category_id INTEGER,
-    user_id INTEGER,
+    category_id BIGINT,
+    user_id BIGINT,
     CONSTRAINT fk_discussions_category_id
         FOREIGN KEY(category_id)
             REFERENCES categories(id) ON DELETE CASCADE,
@@ -52,10 +52,10 @@ CREATE TABLE discussions (
 );
 
 CREATE TABLE posts (
-    id INTEGER PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     content TEXT NOT NULL,
-    user_id INTEGER,
-    discussion_id INTEGER,
+    user_id BIGINT,
+    discussion_id BIGINT,
     CONSTRAINT fk_posts_discussion_id
         FOREIGN KEY(discussion_id)
             REFERENCES discussions(id) ON DELETE CASCADE,
